pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hr-alerter"
version = "0.1.0"
description = "Monitor Polish job boards for HR hiring activity and generate weekly lead reports"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "pypyr>=5.9",
    "requests>=2.28",
    "beautifulsoup4>=4.12",
    "python-dotenv>=1.0",
    "jinja2>=3.1",
]

[project.optional-dependencies]
playwright = ["playwright>=1.40"]

[project.scripts]
hr-alerter = "hr_alerter.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src/hr_alerter"]
